billing service/src/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy
Flask-API
psycopg2
flask-wtf==0.14.3
wtforms==2.3.3
email-validator==1.1.1